# server/api/openapi.yaml
openapi: 3.1.0
info:
  title: doc-thor server API
  description: >
    Core API for doc-thor. Manages documentation projects, build jobs,
    published versions, authentication, and backend discovery.
  version: 1.0.0
servers:
  - url: /api/v1

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
# All endpoints except GET /health and POST /auth/login require a Bearer
# token.  The token is either a session token (returned by /auth/login) or
# an API key (created via POST /auth/apikey).  Both are validated the same
# way; only the raw value is ever sent over the wire — the server stores
# only its SHA-256 hash.
# ---------------------------------------------------------------------------
security:
  - bearerToken: []
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: Session token from POST /auth/login or an API key from POST /auth/apikey.

  # -----------------------------------------------------------------
  # Reusable schemas
  # -----------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: project not found

    # --- Project ---
    Project:
      type: object
      required:
        - id
        - slug
        - name
        - source_url
        - docker_image
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: uint
        slug:
          type: string
        name:
          type: string
        source_url:
          type: string
        docker_image:
          type: string
          description: >
            Docker image the builder will run for this project.  The image
            must follow the builder contract: source repo is mounted
            read-only at /repo, generated output must be written to /output.
          example: doc-thor/builder-mkdocs
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    VCSIntegration:
      type: object
      required:
        - id
        - name
        - provider
        - instance_url
        - enabled
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: uint
        name:
          type: string
          description: Unique identifier for this VCS integration (e.g., "company-gitlab")
          example: company-gitlab
        provider:
          type: string
          enum: [gitlab, github, gitea]
        instance_url:
          type: string
          description: Base URL of the VCS instance
          example: https://gitlab.company.com
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    VCSIntegrationCreate:
      type: object
      required:
        - name
        - provider
        - instance_url
        - access_token
        - webhook_secret
      properties:
        name:
          type: string
          example: company-gitlab
        provider:
          type: string
          enum: [gitlab, github, gitea]
        instance_url:
          type: string
          example: https://gitlab.company.com
        access_token:
          type: string
          description: API token for accessing the VCS platform
        webhook_secret:
          type: string
          description: Secret for validating webhook signatures
        enabled:
          type: boolean
          default: true
    VCSIntegrationUpdate:
      type: object
      properties:
        provider:
          type: string
          enum: [gitlab, github, gitea]
        instance_url:
          type: string
        access_token:
          type: string
        webhook_secret:
          type: string
        enabled:
          type: boolean
    DiscoveryRequest:
      type: object
      required:
        - scope
      properties:
        scope:
          type: string
          description: VCS scope to scan (e.g., GitLab group path)
          example: myteam/docs
    DiscoveredProject:
      type: object
      required:
        - name
        - path
        - clone_url
        - default_branch
        - has_doc_thor
      properties:
        name:
          type: string
        path:
          type: string
          description: Full repository path in VCS
          example: myteam/docs/api-docs
        clone_url:
          type: string
        default_branch:
          type: string
        has_doc_thor:
          type: boolean
        doc_thor_config:
          $ref: "#/components/schemas/DocThorConfig"
    DocThorConfig:
      type: object
      required:
        - slug
        - name
        - docker_image
      properties:
        slug:
          type: string
        name:
          type: string
        docker_image:
          type: string
        branch_mappings:
          type: array
          items:
            $ref: "#/components/schemas/BranchMapping"
    BranchMapping:
      type: object
      required:
        - branch
        - version_tag
      properties:
        branch:
          type: string
          description: Branch/tag pattern (supports glob)
          example: main
        version_tag:
          type: string
          description: Target version tag (supports ${branch} and ${tag} variables)
          example: latest
        auto_publish:
          type: boolean
          default: false
    ImportProjectRequest:
      type: object
      required:
        - integration_name
        - discovered_project
      properties:
        integration_name:
          type: string
        discovered_project:
          $ref: "#/components/schemas/DiscoveredProject"
        branch_mappings:
          type: array
          items:
            $ref: "#/components/schemas/BranchMapping"
        auto_publish:
          type: boolean
          description: Override auto_publish for all branch mappings
        register_webhook:
          type: boolean
          default: false
        callback_url:
          type: string
          description: Webhook callback URL (required if register_webhook is true)

  # -----------------------------------------------------------------
  # Reusable responses
  # -----------------------------------------------------------------
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: unauthorized
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Malformed or invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: database error

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /integrations:
    post:
      summary: Create VCS integration
      operationId: createVCSIntegration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VCSIntegrationCreate"
      responses:
        "201":
          description: Integration created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VCSIntegration"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Integration with this name already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: List all VCS integrations
      operationId: listVCSIntegrations
      responses:
        "200":
          description: Array of VCS integrations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VCSIntegration"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /integrations/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string

    get:
      summary: Get VCS integration details
      operationId: getVCSIntegration
      responses:
        "200":
          description: The requested integration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VCSIntegration"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      summary: Update VCS integration
      operationId: updateVCSIntegration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VCSIntegrationUpdate"
      responses:
        "200":
          description: Updated integration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VCSIntegration"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      summary: Delete VCS integration
      operationId: deleteVCSIntegration
      responses:
        "204":
          description: Integration deleted.
        "400":
          description: Cannot delete - projects are using this integration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /integrations/{name}/test:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string

    post:
      summary: Test VCS integration connection
      operationId: testVCSIntegration
      responses:
        "200":
          description: Connection test result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  error:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # -----------------------------------------------------------------------
  # Project Discovery
  # -----------------------------------------------------------------------
  /integrations/{name}/discover:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
        description: VCS integration name

    post:
      summary: Discover projects in VCS scope
      description: Scans the given VCS scope for projects with .doc-thor.project.yaml
      operationId: discoverProjects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscoveryRequest"
      responses:
        "200":
          description: Discovered projects.
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/DiscoveredProject"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Integration not found.
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /projects/import:
    post:
      summary: Import a discovered project
      description: Creates a project from discovery results and optionally registers webhook
      operationId: importProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportProjectRequest"
      responses:
        "201":
          description: Project imported.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Integration not found.
          $ref: "#/components/responses/NotFound"
        "409":
          description: Project with this slug already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

  # -----------------------------------------------------------------------
  # Webhooks (Public endpoints - no authentication)
  # -----------------------------------------------------------------------
...